Moved all InMemory implementations to inmemory package
