+++ src/main/java/com/training/fnsrv/dao/inmemory/InMemoryIpRouteDao.java
@@ -1,26 +1,27 @@
-package com.training.fnsrv.dao;
+package com.training.fnsrv.dao.inmemory;
 
-import com.training.fnsrv.model.IpInterface;
+import com.training.fnsrv.dao.IpRouteDao;
+import com.training.fnsrv.model.IpRoute;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 import java.util.Optional;
 
 @Component
-public class InMemoryIpInterfaceDao implements IpInterfaceDao {
+public class InMemoryIpRouteDao implements IpRouteDao {
 
     @Override
-    public IpInterface save(IpInterface intf) {
-        return intf;
+    public IpRoute save(IpRoute route) {
+        return route;
     }
 
     @Override
-    public <S extends IpInterface> List<S> saveAll(Iterable<S> intf) {
+    public <S extends IpRoute> List<S> saveAll(Iterable<S> route) {
         return null;
     }
 
     @Override
-    public Optional<IpInterface> findById(Long id) {
+    public Optional<IpRoute> findById(Long id) {
         return null;
     }
 
@@ -30,12 +31,12 @@ public class InMemoryIpInterfaceDao implements IpInterfaceDao {
     }
 
     @Override
-    public Iterable<IpInterface> findAll() {
+    public Iterable<IpRoute> findAll() {
         return null;
     }
 
     @Override
-    public Iterable<IpInterface> findAllById(Iterable<Long> ids) {
+    public Iterable<IpRoute> findAllById(Iterable<Long> ids) {
         return null;
     }
 
@@ -48,10 +49,10 @@ public class InMemoryIpInterfaceDao implements IpInterfaceDao {
     public void deleteById(Long id) { }
 
     @Override
-    public void delete(IpInterface intf) { }
+    public void delete(IpRoute route) { }
 
     @Override
-    public void deleteAll(Iterable<? extends IpInterface> intfs) { }
+    public void deleteAll(Iterable<? extends IpRoute> routes) { }
 
     @Override
     public void deleteAll() { }
