+++ src/main/java/com/training/fnsrv/service/HostService.java
@@ -2,7 +2,7 @@ package com.training.fnsrv.service;
 
 import com.google.gson.Gson;
 import com.google.protobuf.util.JsonFormat;
-import com.training.fnsrv.dao.InMemoryHostDao;
+import com.training.fnsrv.dao.inmemory.InMemoryHostDao;
 import com.training.fnsrv.model.*;
 import com.training.fnsrv.protobuff.gen.HostProto;
 import com.training.fnsrv.task.HostTask;
